Change hat tree shape to 14 levels with 16 bits per level

+++ src/Interfaces/IHatsIdUtilities.sol
@@ -15,10 +15,8 @@
 
 pragma solidity >=0.8.13;
 
-import "./IHatsIdUtilities.sol";
-
 interface IHatsIdUtilities {
-    function buildHatId(uint256 _admin, uint8 _newHat)
+    function buildHatId(uint256 _admin, uint16 _newHat)
         external
         pure
         returns (uint256 id);
